fixing bugs with cascade deletion
