+++ project/routers/users.py
@@ -9,10 +9,10 @@ from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy.orm import Session
 
 from project import database
+from project import models
 from project import utils
 from project.schemas import Token
 from project.users import crud
-from project.users import models
 from project.users import schemas
 
 router = APIRouter(prefix="/users", tags=["Users"])
